Move time tests
